videoProcessingSerivce/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
future==0.18.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
opencv-python-headless==4.8.1.78
platformdirs==4.1.0
pyee==11.1.0
python-ffmpeg==2.0.9
pytube==15.0.0
requests==2.31.0
scenedetect==0.6.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
